vkd3d: Fix formatting warnings.
Extremely annoying and there's a reason why we'd had -Wno-format,
but there were tons of (mostly benign) formatting bugs all over the place.

Native and MinGW builds use different types for things mostly due to
long. There seems to be no good way to have MinGW assume that %zu works since it still
targets msvcrt.dll from prehistoric era, so needed a hack to get around that too.

+++ libs/vkd3d/meta_commands.c
@@ -284,7 +284,7 @@ static HRESULT STDMETHODCALLTYPE d3d12_meta_command_QueryInterface(d3d12_meta_co
 static ULONG STDMETHODCALLTYPE d3d12_meta_command_AddRef(d3d12_meta_command_iface *iface)
 {
     struct d3d12_meta_command *meta_command = impl_from_ID3D12MetaCommand(iface);
-    ULONG refcount = InterlockedIncrement(&meta_command->refcount);
+    unsigned int refcount = InterlockedIncrement(&meta_command->refcount);
 
     TRACE("%p increasing refcount to %u.\n", meta_command, refcount);
 
@@ -302,7 +302,7 @@ static ULONG STDMETHODCALLTYPE d3d12_meta_command_Release(d3d12_meta_command_ifa
 {
     struct d3d12_meta_command *meta_command = impl_from_ID3D12MetaCommand(iface);
     struct d3d12_device *device = meta_command->device;
-    ULONG refcount;
+    unsigned int refcount;
 
     refcount = InterlockedDecrement(&meta_command->refcount);
 
@@ -558,7 +558,7 @@ static HRESULT d3d12_meta_command_create_dstorage(struct d3d12_meta_command *met
 
     if (parameter_size < sizeof(*parameters) || !parameters)
     {
-        FIXME("Invalid parameter set (size = %zu, ptr = %p) for DirectStorage meta command.\n", parameter_size, parameters);
+        FIXME("Invalid parameter set (size = %"PRIuPTR", ptr = %p) for DirectStorage meta command.\n", parameter_size, parameters);
         return E_INVALIDARG;
     }
